テスト用の設定のパッケージを見直し
