+++ sqlmapper-parent/sqlmapper-core/src/test/java/com/github/mygreen/sqlmapper/core/query/auto/SamplePgsqlStoredTest.java
@@ -31,7 +31,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import com.github.mygreen.sqlmapper.core.SqlMapper;
 import com.github.mygreen.sqlmapper.core.annotation.In;
 import com.github.mygreen.sqlmapper.core.annotation.Out;
-import com.github.mygreen.sqlmapper.core.testdata.PgsqlTestConfig;
+import com.github.mygreen.sqlmapper.core.test.config.PgsqlTestConfig;
 
 import lombok.Data;
