+++ sqlmapper-parent/sqlmapper-core/src/test/java/com/github/mygreen/sqlmapper/core/test/entity/MBusinessEstablishment.java
@@ -1,4 +1,4 @@
-package com.github.mygreen.sqlmapper.core.testdata;
+package com.github.mygreen.sqlmapper.core.test.entity;
 
 import com.github.mygreen.sqlmapper.metamodel.EntityPathBase;
 import com.github.mygreen.sqlmapper.metamodel.NumberPath;
