+++ sqlmapper-parent/sqlmapper-core/src/test/java/com/github/mygreen/sqlmapper/core/test/entity/SampleSequence1.java
@@ -1,4 +1,4 @@
-package com.github.mygreen.sqlmapper.core.testdata;
+package com.github.mygreen.sqlmapper.core.test.entity;
 
 import com.github.mygreen.sqlmapper.core.annotation.Entity;
 import com.github.mygreen.sqlmapper.core.annotation.GeneratedValue;
@@ -8,16 +8,16 @@ import com.github.mygreen.sqlmapper.core.annotation.GeneratedValue.GenerationTyp
 import lombok.Data;
 
 /**
- * 自動生成の主キーテーブルで作成する。
+ * 自動生成の主キーが２つ
  *
  *
  */
 @Data
 @Entity
-public class SampleTableId1 {
+public class SampleSequence1 {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.TABLE)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
     public Long id;
 
     public String value;
