+++ sqlmapper-parent/sqlmapper-core/src/test/java/com/github/mygreen/sqlmapper/core/test/entity/SampleTableId1.java
@@ -1,4 +1,4 @@
-package com.github.mygreen.sqlmapper.core.testdata;
+package com.github.mygreen.sqlmapper.core.test.entity;
 
 import com.github.mygreen.sqlmapper.core.annotation.Entity;
 import com.github.mygreen.sqlmapper.core.annotation.GeneratedValue;
@@ -8,21 +8,17 @@ import com.github.mygreen.sqlmapper.core.annotation.GeneratedValue.GenerationTyp
 import lombok.Data;
 
 /**
- * 自動生成の主キーが２つ
+ * 自動生成の主キーテーブルで作成する。
  *
  *
  */
 @Data
 @Entity
-public class SampleIdentity2 {
+public class SampleTableId1 {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Long id1;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY, format="00000")
-    public String id2;
+    @GeneratedValue(strategy = GenerationType.TABLE)
+    public Long id;
 
     public String value;
