+++ sqlmapper-parent/sqlmapper-core/src/test/java/com/github/mygreen/sqlmapper/core/test/package-info.java
@@ -1,4 +1,4 @@
 /**
  * テスト用の設定やデータを提供します。
  */
-package com.github.mygreen.sqlmapper.core.testdata;
+package com.github.mygreen.sqlmapper.core.test;
