+++ sqlmapper-parent/sqlmapper-core/src/test/java/com/github/mygreen/sqlmapper/core/where/metamodel/MetamodelWhereBuilderTest.java
@@ -17,13 +17,13 @@ import com.github.mygreen.sqlmapper.core.dialect.Dialect;
 import com.github.mygreen.sqlmapper.core.meta.EntityMeta;
 import com.github.mygreen.sqlmapper.core.meta.EntityMetaFactory;
 import com.github.mygreen.sqlmapper.core.query.TableNameResolver;
-import com.github.mygreen.sqlmapper.core.testdata.Customer;
-import com.github.mygreen.sqlmapper.core.testdata.EmbeddedEntity;
-import com.github.mygreen.sqlmapper.core.testdata.EntityChild;
-import com.github.mygreen.sqlmapper.core.testdata.MCustomer;
-import com.github.mygreen.sqlmapper.core.testdata.MEmbeddedEntity;
-import com.github.mygreen.sqlmapper.core.testdata.MEntityChild;
-import com.github.mygreen.sqlmapper.core.testdata.TestConfig;
+import com.github.mygreen.sqlmapper.core.test.config.TestConfig;
+import com.github.mygreen.sqlmapper.core.test.entity.Customer;
+import com.github.mygreen.sqlmapper.core.test.entity.EmbeddedEntity;
+import com.github.mygreen.sqlmapper.core.test.entity.EntityChild;
+import com.github.mygreen.sqlmapper.core.test.entity.MCustomer;
+import com.github.mygreen.sqlmapper.core.test.entity.MEmbeddedEntity;
+import com.github.mygreen.sqlmapper.core.test.entity.MEntityChild;
 import com.github.mygreen.sqlmapper.metamodel.Predicate;
 import com.github.mygreen.sqlmapper.metamodel.support.SubQueryHelper;
